[Move `@kbn/config-schema` to server] `home`
