+++ src/plugins/home/server/config.ts
