+++ src/plugins/home/server/index.ts
@@ -23,7 +23,7 @@ export type {
   ScopedTutorialContextFactory,
 } from './services';
 import { PluginInitializerContext, PluginConfigDescriptor } from '@kbn/core/server';
-import { configSchema, ConfigSchema } from '../config';
+import { configSchema, ConfigSchema } from './config';
 
 export const config: PluginConfigDescriptor<ConfigSchema> = {
   exposeToBrowser: {
